Reformatted code to reuse use draft button
